+++ frontend/buildings.app.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDom from "react-dom";
 import { BrowserRouter, Route } from "react-router-dom";
-import Navbar from "./src/components/Navbar";
+import TopMenu from "./src/components/library/TopMenu";
 import BuildingsIndex from "./src/components/buildings/BuildingsIndexContainer";
 
 const Apartments = () => <h1>Apartments</h1>;
@@ -9,7 +9,7 @@ const Apartments = () => <h1>Apartments</h1>;
 const app = (
   <BrowserRouter>
     <div>
-      <Navbar/>
+      <TopMenu/>
       <Route exact path="/app/buildings" component={BuildingsIndex} />
       <Route
         exact
